requirements.txt
Django==1.5.1
Scrapy==0.16.1
South==0.7.6
Twisted==12.2.0
argparse==1.2.1
beautifulsoup4==4.1.3
django-tastypie==0.9.15
guppy==0.1.10
html5lib==0.95
lxml==3.0.1
mimeparse==0.1.3
psycopg2==2.5
pyOpenSSL==0.13
python-dateutil==2.1
requests==0.14.2
six==1.3.0
w3lib==1.2
wsgiref==0.1.2
zope.interface==4.0.1
